requirements.txt
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.12
django-allauth==0.41.0
django-autoslug==1.9.7
django-crispy-forms==1.9.1
idna==2.9
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
